Added animation when boss is defeated

+++ main.ts
@@ -1,27 +1,26 @@
 /**
- * De variabele worden op de beginwaarden ingesteld.
+ * Maakt een kogel aan als de speler op A+B drukt.
  */
 /**
  * Beweegt de speler naar links of rechts.
  */
 /**
- * Maakt een kogel aan als de speler op A+B drukt.
+ * De variabele worden op de beginwaarden ingesteld.
  */
 /**
- * Genereert een getal tussen de 0 en 3, als dit 2 is wordt er een enemtShipSmall gespawned.
+ * Start een animatie als de eindbaas verslagen is.
+ */
+/**
+ * Haalt de particles vande animatie weg als ze de rand aanraken.
+ */
+/**
+ * Genereert een getal tussen de 1 en 3, als dit 2 is wordt er een enemtShipSmall gespawned.
  */
 /**
  * Ik gebruik hier expres geen else if statements, omdat er soms meerdere voorwaarden gecontroleerd moeten worden.
  */
 /**
- * - Update de score
- * 
- * - Hoe meer punten de speler heeft gehaald, hoe sneller het spel gaat
- * 
- * - Als de speler meer dan 50 punten heeft gehaald begint de eindbaas.
- */
-/**
- * Verplaatst de enemyShipSmall naar beneden elke (0.3*eventInterval) seconden.
+ * Spawned een eindbaas.
  */
 /**
  * Kijkt of er een enemyShipSmall de onderste rand aanraakt. Als dit zo is wordt het schip verwijderd.
@@ -30,10 +29,11 @@
  * Kijkt of er een enemyShipSmall de speler aanraakt. Als dit zo is verliest de speler een leven.
  */
 /**
- * Kijkt of een enemyShipSmall de kogel van de speler aanraakt. Als dit zo is worden de enemyShipSmall en kogel verwijderd en krijgt de speler 5 punten.
- */
-/**
- * Als de eindbaas verslagen is krijgt de speler 15 punten per overgebleven leven.
+ * - Update de score
+ * 
+ * - Hoe meer punten de speler heeft gehaald, hoe sneller het spel gaat
+ * 
+ * - Als de speler meer dan 50 punten heeft gehaald begint de eindbaas.
  */
 input.onButtonPressed(Button.A, function () {
     playerSprite.move(-1)
@@ -41,35 +41,67 @@ input.onButtonPressed(Button.A, function () {
         score += 1
     }
 })
-/**
- * Kijkt of dingen elkaar aanraken en wat er dan moet gebeuren.
- */
 function spawnEnemyShipLarge () {
     enemyShipLargeBodyHeart = game.createSprite(2, 0)
     enemyShipLargeBodyLeft = game.createSprite(1, 0)
     enemyShipLargeBodyFront = game.createSprite(2, 1)
     enemyShipLargeBodyRight = game.createSprite(3, 0)
-    while (enemyShipLargeBodyHeart.isTouchingEdge()) {
-        enemyShipLargeShootingNumber = 5
-        if (enemyShipLargeShootingNumber == 5) {
-            enemyShipLargeBullet = game.createSprite(2, 1)
-            for (let index = 0; index < 3; index++) {
-                basic.pause(enemyShipLargeBulletInterval)
-                enemyShipLargeBullet.change(LedSpriteProperty.Y, 1)
-            }
+    while (enemyShipLargeBodyHeart.isTouchingEdge() && bossIsDefeated < 1) {
+        basic.pause(enemyShipLargeBulletInterval)
+        enemyShipLargeBullet = game.createSprite(2, 1)
+        for (let index = 0; index < 3; index++) {
+            basic.pause(eventInterval)
+            enemyShipLargeBullet.change(LedSpriteProperty.Y, 1)
         }
-        basic.pause(eventInterval)
+        basic.pause(200)
         enemyShipLargeBullet.delete()
     }
 }
+function gameWonAnimation () {
+    enemyShipLargeBullet.delete()
+    ship = game.createSprite(2, 0)
+    ship.set(LedSpriteProperty.Blink, 100)
+    list = []
+    basic.pause(1000)
+    ship.delete()
+    explosionParticleLeft = game.createSprite(1, 0)
+    explosionParticleLeft.set(LedSpriteProperty.Direction, 270)
+    list.push(explosionParticleLeft)
+    explosionParticleDownLeft = game.createSprite(1, 1)
+    explosionParticleDownLeft.set(LedSpriteProperty.Direction, 225)
+    list.push(explosionParticleDownLeft)
+    explosionParticleRight = game.createSprite(3, 0)
+    explosionParticleRight.set(LedSpriteProperty.Direction, 90)
+    list.push(explosionParticleRight)
+    explosionParticleDownRight = game.createSprite(3, 1)
+    explosionParticleDownRight.set(LedSpriteProperty.Direction, 135)
+    list.push(explosionParticleDownRight)
+    explosionParticleDown = game.createSprite(2, 1)
+    explosionParticleDown.set(LedSpriteProperty.Direction, 180)
+    list.push(explosionParticleDown)
+    basic.pause(100)
+    for (let index = 0; index < 4; index++) {
+        basic.pause(200)
+        for (let value of list) {
+            value.change(LedSpriteProperty.Brightness, -75)
+            value.move(1)
+        }
+    }
+    playerSprite.delete()
+    basic.pause(500)
+    basic.showString("WON SCORE")
+    basic.showNumber(score)
+}
 input.onButtonPressed(Button.AB, function () {
-    playerBullet = game.createSprite(playerSprite.get(LedSpriteProperty.X), 3)
-    playerBullet.set(LedSpriteProperty.Blink, 350)
-    while (playerBullet) {
-        basic.pause(eventInterval)
-        playerBullet.change(LedSpriteProperty.Y, -1)
-        if (playerBullet.isDeleted()) {
-            break;
+    if (bossIsDefeated < 1) {
+        playerBullet = game.createSprite(playerSprite.get(LedSpriteProperty.X), 3)
+        playerBullet.set(LedSpriteProperty.Blink, 350)
+        while (playerBullet) {
+            basic.pause(eventInterval)
+            playerBullet.change(LedSpriteProperty.Y, -1)
+            if (playerBullet.isDeleted()) {
+                break;
+            }
         }
     }
 })
@@ -79,14 +111,23 @@ input.onButtonPressed(Button.B, function () {
         score += 1
     }
 })
+let bossBattleHasStarted = 0
+let enemyShipSmallList: game.LedSprite[] = []
 let randomNumber = 0
 let playerBullet: game.LedSprite = null
+let explosionParticleDown: game.LedSprite = null
+let explosionParticleDownRight: game.LedSprite = null
+let explosionParticleRight: game.LedSprite = null
+let explosionParticleDownLeft: game.LedSprite = null
+let explosionParticleLeft: game.LedSprite = null
+let list: game.LedSprite[] = []
+let ship: game.LedSprite = null
 let enemyShipLargeBullet: game.LedSprite = null
-let enemyShipLargeShootingNumber = 0
 let enemyShipLargeBodyRight: game.LedSprite = null
 let enemyShipLargeBodyFront: game.LedSprite = null
 let enemyShipLargeBodyLeft: game.LedSprite = null
 let enemyShipLargeBodyHeart: game.LedSprite = null
+let bossIsDefeated = 0
 let maxScore = 0
 let enemyShipLargeBulletInterval = 0
 let playerSprite: game.LedSprite = null
@@ -97,18 +138,28 @@ game.setLife(3)
 score = 0
 let lives = 3
 playerSprite = game.createSprite(2, 4)
-playerSprite.set(LedSpriteProperty.Brightness, 175)
+playerSprite.set(LedSpriteProperty.Brightness, 130)
 eventInterval = 1000
-let enemyShipSmallList: game.LedSprite[] = []
-let bossBattleHasStarted = 0
 enemyShipLargeBulletInterval = 400
 maxScore = 50
+bossIsDefeated = 0
+basic.forever(function () {
+    for (let value2 of list) {
+        if (value2.get(LedSpriteProperty.X) == 0 || value2.get(LedSpriteProperty.X) == 4 || value2.get(LedSpriteProperty.Y) == 4) {
+            value2.delete()
+            basic.pause(200)
+        }
+    }
+})
 /**
- * Spawned een eindbaas.
+ * Kijkt of het gevecht met de eindbaas begonnen is en als dit zo is haalt hij de oude kogel van het schip weg.
+ */
+/**
+ * Kijkt of dingen elkaar aanraken en wat er dan moet gebeuren.
  */
 basic.forever(function () {
     while (score < maxScore && game.isRunning()) {
-        randomNumber = randint(0, 3)
+        randomNumber = randint(1, 3)
         if (randomNumber == 2) {
             if (enemyShipSmallList.length < 5) {
                 enemyShipSmallList.unshift(game.createSprite(randint(0, 4), 0))
@@ -118,16 +169,34 @@ basic.forever(function () {
     }
 })
 basic.forever(function () {
+    if (bossIsDefeated == 1) {
+        enemyShipLargeBullet.delete()
+    }
+})
+/**
+ * Kijkt of een enemyShipSmall de kogel van de speler aanraakt. Als dit zo is worden de enemyShipSmall en kogel verwijderd en krijgt de speler 5 punten.
+ */
+basic.forever(function () {
+    for (let value7 of enemyShipSmallList) {
+        if (playerBullet && value7.isTouching(playerBullet)) {
+            playerBullet.delete()
+            enemyShipSmallList.removeAt(enemyShipSmallList.indexOf(value7))
+            value7.delete()
+            score += 5
+        }
+    }
+})
+basic.forever(function () {
     if (enemyShipLargeBodyHeart && playerBullet) {
         if (playerBullet.isTouching(enemyShipLargeBodyLeft)) {
             enemyShipLargeBodyLeft.delete()
             playerBullet.delete()
-            enemyShipLargeBulletInterval += 100
+            enemyShipLargeBulletInterval += 300
         }
         if (playerBullet.isTouching(enemyShipLargeBodyRight)) {
             enemyShipLargeBodyRight.delete()
             playerBullet.delete()
-            enemyShipLargeBulletInterval += 100
+            enemyShipLargeBulletInterval += 300
         }
         if (playerBullet.isTouching(enemyShipLargeBodyFront)) {
             enemyShipLargeBodyFront.delete()
@@ -135,19 +204,25 @@ basic.forever(function () {
             enemyShipLargeBulletInterval += 100
         }
         if (playerBullet.isTouching(enemyShipLargeBodyHeart)) {
+            enemyShipLargeBullet.delete()
+            bossIsDefeated = 1
             enemyShipLargeBodyHeart.delete()
+            enemyShipLargeBodyLeft.delete()
+            enemyShipLargeBodyRight.delete()
             playerBullet.delete()
+            score += lives * 15
+            gameWonAnimation()
         }
-        if (playerBullet.isTouching(enemyShipLargeBullet)) {
+        if (enemyShipLargeBullet && playerBullet.isTouching(enemyShipLargeBullet)) {
             enemyShipLargeBullet.delete()
             playerBullet.delete()
         }
     }
     if (enemyShipLargeBullet) {
         if (playerSprite.isTouching(enemyShipLargeBullet)) {
+            enemyShipLargeBullet.delete()
             lives += -1
             game.removeLife(1)
-            enemyShipLargeBullet.delete()
         }
     }
     if (playerBullet) {
@@ -158,56 +233,43 @@ basic.forever(function () {
         }
     }
 })
+/**
+ * Verplaatst de enemyShipSmall naar beneden elke (0.3*eventInterval) seconden.
+ */
+basic.forever(function () {
+    for (let value4 of enemyShipSmallList) {
+        basic.pause(eventInterval * 0.3)
+        value4.change(LedSpriteProperty.Y, 1)
+        value4.set(LedSpriteProperty.Brightness, 150)
+    }
+})
+basic.forever(function () {
+    for (let value5 of enemyShipSmallList) {
+        if (value5.get(LedSpriteProperty.Y) == 4) {
+            basic.pause(200)
+            enemyShipSmallList.removeAt(enemyShipSmallList.indexOf(value5))
+            value5.delete()
+        }
+    }
+})
+basic.forever(function () {
+    for (let value6 of enemyShipSmallList) {
+        if (value6.isTouching(playerSprite)) {
+            game.removeLife(1)
+            value6.delete()
+            enemyShipSmallList.removeAt(enemyShipSmallList.indexOf(value6))
+        }
+    }
+})
 basic.forever(function () {
     game.setScore(score)
     eventInterval = 1000 - score * 10
     if (score >= maxScore && bossBattleHasStarted < 1) {
-        for (let value of enemyShipSmallList) {
+        for (let value3 of enemyShipSmallList) {
             enemyShipSmallList = []
-            value.delete()
+            value3.delete()
         }
         spawnEnemyShipLarge()
         bossBattleHasStarted += 1
-    }
-})
-basic.forever(function () {
-    for (let value of enemyShipSmallList) {
-        basic.pause(eventInterval * 0.3)
-        value.change(LedSpriteProperty.Y, 1)
-        value.set(LedSpriteProperty.Brightness, 100)
-    }
-})
-basic.forever(function () {
-    for (let value of enemyShipSmallList) {
-        if (value.get(LedSpriteProperty.Y) == 4) {
-            basic.pause(200)
-            enemyShipSmallList.removeAt(enemyShipSmallList.indexOf(value))
-            value.delete()
-        }
-    }
-})
-basic.forever(function () {
-    for (let value of enemyShipSmallList) {
-        if (value.isTouching(playerSprite)) {
-            game.removeLife(1)
-            value.delete()
-            enemyShipSmallList.removeAt(enemyShipSmallList.indexOf(value))
-        }
-    }
-})
-basic.forever(function () {
-    for (let value of enemyShipSmallList) {
-        if (playerBullet && value.isTouching(playerBullet)) {
-            value.delete()
-            enemyShipSmallList.removeAt(enemyShipSmallList.indexOf(value))
-            playerBullet.delete()
-            score += 5
-        }
-    }
-})
-basic.forever(function () {
-    if (bossBattleHasStarted == 1 && enemyShipLargeBodyHeart.isDeleted()) {
-        score += lives * 15
-        game.gameOver()
     }
 })
